finish multiply and divide tests

+++ Calculator/Calculator.cs
@@ -22,7 +22,7 @@ public class Calculator
         return a * b;
     }
 
-    public int Divide(int a, int b)
+    public double Divide(double a, double b)
     {
         if (b == 0)
         {
